requirements.txt
asttokens==3.0.0
cramjam==2.9.1
cypari==2.5.5
decorator==5.1.1
executing==2.2.0
FXrays==1.3.6
ipython==8.32.0
jedi==0.19.2
knot_floer_homology==1.2.2
low_index==1.2.1
matplotlib-inline==0.1.7
networkx==3.4.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
plink==2.4.5
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pypng==0.20220715.0
python-snappy==0.7.3
PyX==0.16
snappy_manifolds==1.2.1
spherogram==2.3
stack-data==0.6.3
tkinter-gl==1.0
traitlets==5.14.3
wcwidth==0.2.13
